Check that we won't override other's changes in publish.sh

+++ publish.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
+BOLD="\033[1m"
+RED="\033[38;5;9m"
+RESET="\033[0m"
 set -x -e -o pipefail # Show commands executed, exit on error (including subcommands)
 
 # Check if hugo is installed
@@ -8,12 +11,19 @@ if ! command -v hugo >/dev/null ; then
     exit 1
 fi
 
+# We want to check that the remote doesn't have any changes we will override
+git fetch origin master
+if ! git merge-base --is-ancestor origin/master master ; then
+    ( set +x; echo -e "${BOLD}${RED}* ERROR: There are changes on the remote!\n* Please pull and merge with the remote before publishing. Otherwise you might override other people's changes${RESET}" )
+    exit 1
+fi
+
+
 # At the end of the script, including if it fails, we want to restore to the master branch
 cleanup() {
     git checkout master
 }
 trap cleanup EXIT
-
 echo "Deleting old publication"
 rm -rf public
